# openapi.yaml
openapi: 3.1.0
info:
  title: Hampshire Bin Collection API
  version: 1.0.0
  description: |
    API for accessing bin collection schedules across Hampshire local councils.
    
    ## Authentication
    
    Public endpoints require an API key passed via the `X-API-Key` header.
    Admin endpoints require a JWT bearer token.
    
    ## Rate Limiting
    
    Default rate limit is 60 requests per minute per API key.
    Rate limit headers are included in all responses.
    
    ## Errors
    
    All errors follow a consistent format with `code`, `message`, and optional `details`.
  contact:
    name: Hampshire Bins Platform
    email: support@hampshire-bins.example.com
  license:
    name: MIT
    identifier: MIT
servers:
  - url: https://api.hampshire-bins.example.com
    description: Production
  - url: https://staging-api.hampshire-bins.example.com
    description: Staging
  - url: http://localhost:3000
    description: Local development
...
security:
  - ApiKeyAuth: []
paths:
  /v1/councils:
    get:
      operationId: listCouncils
      summary: List all councils
      description: Returns a list of all supported Hampshire councils with adapter status.
      tags: [Councils]
      parameters:
        - name: status
          in: query
          description: Filter by adapter status
          schema:
            type: string
            enum: [active, beta, development, disabled, unsupported]
      responses:
        '200':
          description: List of councils
          headers:
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Council'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/councils/{councilId}:
    get:
      operationId: getCouncil
      summary: Get council details
      description: Returns detailed information about a specific council.
      tags: [Councils]
      parameters:
        - name: councilId
          in: path
          required: true
          description: Council identifier
          schema:
            type: string
            pattern: '^[a-z][a-z0-9-]*[a-z0-9]$'
            example: basingstoke
      responses:
        '200':
          description: Council details
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/CouncilDetail'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API key for public endpoints. Obtain from the platform administrators.
        Format: `hb_live_xxxxxxxxxxxxxxxxxxxx` (32 characters after prefix)
    
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token for admin endpoints. Obtained via OAuth 2.0 flow.
        Token must include `admin` scope.
  headers:
    X-RateLimit-Limit:
      description: Request limit per minute
      schema:
        type: integer
        example: 60
    X-RateLimit-Remaining:
      description: Remaining requests in current window
      schema:
        type: integer
        example: 45
    X-RateLimit-Reset:
      description: Unix timestamp when rate limit resets
      schema:
        type: integer
        example: 1711364400
  schemas:
    Council:
      type: object
      required: [councilId, councilName, adapterStatus, lookupMethod, upstreamRiskLevel]
      properties:
        councilId:
          type: string
          description: Unique council identifier (kebab-case)
          enum:
            - basingstoke-deane
            - east-hampshire
            - eastleigh
            - fareham
            - gosport
            - hart
            - havant
            - new-forest
            - portsmouth
            - rushmoor
            - southampton
            - test-valley
            - winchester
          example: eastleigh
        councilName:
          type: string
          description: Official council name
          example: Eastleigh Borough Council
        adapterStatus:
          type: string
          enum: [implemented, postponed, stub, disabled]
          description: Current adapter implementation status
        lookupMethod:
          type: string
          enum: [api, hidden_json, html_form, pdf_calendar, browser_automation, browser_json, unknown, unsupported]
          description: Method used by adapter to acquire data
        upstreamRiskLevel:
          type: string
          enum: [low, medium, high, critical]
          description: Risk level of upstream data source (brittleness, bot protection, etc.)
        killSwitchActive:
          type: boolean
          description: Kill switch status (admin-only field)
        lastHealthCheck:
          type: string
          format: date-time
          description: Last health check timestamp (admin-only field)
        currentConfidence:
          type: number
          format: float
          minimum: 0.0
          maximum: 1.0
          description: Current confidence score (admin-only field)
    CouncilDetail:
      allOf:
        - $ref: '#/components/schemas/Council'
        - type: object
          properties:
            website:
              type: string
              format: uri
              description: Council website URL
            binInfoUrl:
              type: string
              format: uri
              description: URL to bin collection lookup page
            supportedServiceTypes:
              type: array
              items:
                type: string
              description: Service types supported by this council
            capabilities:
              type: object
              properties:
                supports_address_lookup:
                  type: boolean
                supports_collection_services:
                  type: boolean
                supports_collection_events:
                  type: boolean
                provides_uprn:
                  type: boolean
                max_event_range_days:
                  type: integer
            limitations:
              type: array
              items:
                type: string
              description: Known limitations of this adapter
            is_production_ready:
              type: boolean
            last_updated:
              type: string
              format: date
            health:
              type: object
              properties:
                status:
                  type: string
                  enum: [healthy, degraded, unavailable]
                last_success:
                  type: string
                  format: date-time
                success_rate_24h:
                  type: number
                  format: float
                avg_response_time_ms:
                  type: number
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code
          example: INVALID_POSTCODE
        message:
          type: string
          description: Human-readable error message
          example: The postcode format is invalid
        details:
          type: object
          description: Additional error details
          additionalProperties: true
        correlationId:
          type: string
          description: Request correlation ID for support
  responses:
...
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: UNAUTHORIZED
            message: Invalid or missing API key
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: NOT_FOUND
            message: The requested resource was not found
    RateLimitExceeded:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/X-RateLimit-Limit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/X-RateLimit-Remaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/X-RateLimit-Reset'
        Retry-After:
          description: Seconds until rate limit resets
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: RATE_LIMIT_EXCEEDED
            message: Rate limit exceeded. Try again later.
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: INTERNAL_ERROR
            message: An unexpected error occurred